backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pymongo==4.8.0
snowflake-connector-python==3.11.0
requests==2.32.3
pdfplumber==0.11.4
openpyxl==3.1.5
PyMySQL==1.1.1

